move 5.0 library module into 5.3
