+++ instrumentation/spring/spring-boot-autoconfigure/README.md
@@ -2,7 +2,7 @@
 
 Auto-configures OpenTelemetry instrumentation for [spring-web](../spring-web/spring-web-3.1/library)
 , [spring-webmvc](../spring-webmvc/spring-webmvc-5.3/library),
-and [spring-webflux](../spring-webflux/spring-webflux-5.0/library). Leverages Spring Aspect Oriented
+and [spring-webflux](../spring-webflux/spring-webflux-5.3/library). Leverages Spring Aspect Oriented
 Programming,
 dependency injection, and bean post-processing to trace spring applications. To include all features
 listed below use the [opentelemetry-spring-boot-starter](../starters/spring-boot-starter/README.md).
@@ -168,7 +168,16 @@ to learn more about the OpenTelemetry Spring WebMVC instrumentation.
 
 #### Spring WebFlux Auto Configuration
 
-Provides auto-configurations for the OpenTelemetry WebClient ExchangeFilter defined in [opentelemetry-spring-webflux-5.0](../spring-webflux/spring-webflux-5.0). This auto-configuration instruments all outgoing http requests sent using Spring's WebClient and WebClient Builder beans by applying a bean post processor. This feature is supported for spring webflux versions 5.0+ and can be disabled by adding `opentelemetry.trace.httpclients.enabled=false` to your `resources/applications.properties` file. [Spring Web-Flux - WebClient Span](#spring-web-flux---webclient-span) showcases a sample span generated by the WebClientFilter. Check out [opentelemetry-spring-webflux-5.0](../spring-webflux/spring-webflux-5.0) to learn more about the OpenTelemetry WebClientFilter.
+Provides auto-configurations for the OpenTelemetry WebClient ExchangeFilter defined
+in [opentelemetry-spring-webflux-5.3](../spring-webflux/spring-webflux-5.3). This auto-configuration
+instruments all outgoing http requests sent using Spring's WebClient and WebClient Builder beans by
+applying a bean post processor. This feature is supported for spring webflux versions 5.0+ and can
+be disabled by adding `opentelemetry.trace.httpclients.enabled=false` to
+your `resources/applications.properties`
+file. [Spring Web-Flux - WebClient Span](#spring-web-flux---webclient-span) showcases a sample span
+generated by the WebClientFilter. Check
+out [opentelemetry-spring-webflux-5.3](../spring-webflux/spring-webflux-5.3) to learn more about the
+OpenTelemetry WebClientFilter.
 
 #### Manual Instrumentation Support - @WithSpan
