+++ instrumentation/spring/spring-boot-autoconfigure/build.gradle.kts
@@ -21,7 +21,7 @@ dependencies {
   implementation(project(":instrumentation:spring:spring-webmvc:spring-webmvc-5.3:library"))
   implementation(project(":instrumentation:spring:spring-webmvc:spring-webmvc-6.0:library"))
   compileOnly("jakarta.servlet:jakarta.servlet-api:5.0.0")
-  implementation(project(":instrumentation:spring:spring-webflux:spring-webflux-5.0:library"))
+  implementation(project(":instrumentation:spring:spring-webflux:spring-webflux-5.3:library"))
   implementation(project(":instrumentation:micrometer:micrometer-1.5:library"))
 
   compileOnly("org.springframework.kafka:spring-kafka:2.9.0")
