+++ instrumentation/spring/spring-webflux/spring-webflux-5.0/javaagent/build.gradle.kts
@@ -39,7 +39,7 @@ muzzle {
 }
 
 dependencies {
-  implementation(project(":instrumentation:spring:spring-webflux:spring-webflux-5.0:library"))
+  implementation(project(":instrumentation:spring:spring-webflux:spring-webflux-5.3:library"))
 
   compileOnly("org.springframework:spring-webflux:5.0.0.RELEASE")
   compileOnly("io.projectreactor.ipc:reactor-netty:0.7.0.RELEASE")
@@ -51,7 +51,7 @@ dependencies {
   testInstrumentation(project(":instrumentation:reactor:reactor-3.1:javaagent"))
   testInstrumentation(project(":instrumentation:reactor:reactor-netty:reactor-netty-1.0:javaagent"))
 
-  testImplementation(project(":instrumentation:spring:spring-webflux:spring-webflux-5.0:testing"))
+  testImplementation(project(":instrumentation:spring:spring-webflux:spring-webflux-5.3:testing"))
 
   testLibrary("org.springframework.boot:spring-boot-starter-webflux:2.0.0.RELEASE")
   testLibrary("org.springframework.boot:spring-boot-starter-test:2.0.0.RELEASE")
