+++ instrumentation/spring/spring-webflux/spring-webflux-5.3/library/build.gradle.kts
@@ -7,7 +7,9 @@ dependencies {
 
   implementation(project(":instrumentation:reactor:reactor-3.1:library"))
 
-  testImplementation(project(":instrumentation:spring:spring-webflux:spring-webflux-5.0:testing"))
+  compileOnly("io.projectreactor.ipc:reactor-netty:0.7.0.RELEASE")
+
+  testImplementation(project(":instrumentation:spring:spring-webflux:spring-webflux-5.3:testing"))
 
   testLibrary("org.springframework.boot:spring-boot-starter-webflux:2.4.0")
   testLibrary("org.springframework.boot:spring-boot-starter-test:2.4.0")
