+++ settings.gradle.kts
@@ -473,8 +473,7 @@ hideFromDependabot(":instrumentation:spring:spring-webmvc:spring-webmvc-6.0:libr
 hideFromDependabot(":instrumentation:spring:spring-webmvc:spring-webmvc-common:javaagent")
 hideFromDependabot(":instrumentation:spring:spring-webmvc:spring-webmvc-common:testing")
 hideFromDependabot(":instrumentation:spring:spring-webflux:spring-webflux-5.0:javaagent")
-hideFromDependabot(":instrumentation:spring:spring-webflux:spring-webflux-5.0:library")
-hideFromDependabot(":instrumentation:spring:spring-webflux:spring-webflux-5.0:testing")
+hideFromDependabot(":instrumentation:spring:spring-webflux:spring-webflux-5.3:testing")
 hideFromDependabot(":instrumentation:spring:spring-webflux:spring-webflux-5.3:library")
 hideFromDependabot(":instrumentation:spring:spring-ws-2.0:javaagent")
 hideFromDependabot(":instrumentation:spring:spring-boot-autoconfigure")
